+++ src/Universum.hs
@@ -31,15 +31,13 @@ module Universum
        , module Universum.TypeOps
        , module Universum.VarArg
 
-         -- * Safe functions
-       , module Safe
-
          -- * Lenses
        , module Lens.Micro
        , module Lens.Micro.Mtl
        ) where
 
 import Universum.Applicative
+import Universum.Base
 import Universum.Bool
 import Universum.Container
 import Universum.Debug
@@ -56,13 +54,6 @@ import Universum.Print
 import Universum.String
 import Universum.TypeOps
 import Universum.VarArg
-
-import Universum.Base hiding (error, show, showFloat, showList, showSigned, showSignedFloat,
-                       showsPrec, undefined)
-
--- Maybe'ized version of partial functions
-import Safe (atDef, atMay, foldl1May, foldr1May, headDef, headMay, initDef, initMay, initSafe,
-             lastDef, lastMay, tailDef, tailMay, tailSafe)
 
 -- Lenses
 import Lens.Micro (Lens, Lens', Traversal, Traversal', over, set, (%~), (&), (.~), (<&>), (^.),